+++ railties/lib/rails/application/configuration.rb
@@ -264,6 +264,7 @@ module Rails
 
           if respond_to?(:active_record)
             active_record.run_commit_callbacks_on_first_saved_instances_in_transaction = false
+            active_record.allow_deprecated_singular_associations_name = false
           end
 
           if respond_to?(:action_dispatch)